+++ internal/dinosql/testdata/ondeck/db.go
@@ -21,7 +21,7 @@ type City struct {
 }
 
 type Venue struct {
-	ID              int
+	ID              int32
 	Status          Status
 	Slug            string
 	Name            string
@@ -104,9 +104,9 @@ type CreateVenueParams struct {
 	Tags            []string
 }
 
-func (q *Queries) CreateVenue(ctx context.Context, arg CreateVenueParams) (int, error) {
+func (q *Queries) CreateVenue(ctx context.Context, arg CreateVenueParams) (int32, error) {
 	row := q.db.QueryRowContext(ctx, createVenue, arg.Slug, arg.Name, arg.City, arg.SpotifyPlaylist, arg.Status, pq.Array(arg.Tags))
-	var id int
+	var id int32
 	err := row.Scan(&id)
 	return id, err
 }
@@ -239,9 +239,9 @@ type UpdateVenueNameParams struct {
 	Name string
 }
 
-func (q *Queries) UpdateVenueName(ctx context.Context, arg UpdateVenueNameParams) (int, error) {
+func (q *Queries) UpdateVenueName(ctx context.Context, arg UpdateVenueNameParams) (int32, error) {
 	row := q.db.QueryRowContext(ctx, updateVenueName, arg.Slug, arg.Name)
-	var id int
+	var id int32
 	err := row.Scan(&id)
 	return id, err
 }
@@ -257,7 +257,7 @@ ORDER BY 1
 
 type VenueCountByCityRow struct {
 	City  string
-	Count int
+	Count int64
 }
 
 func (q *Queries) VenueCountByCity(ctx context.Context) ([]VenueCountByCityRow, error) {
